refactor(python): split collection.rs (887L) into focused sub-modules
Split the monolithic PyO3 collection module into 5 files under
collection/, all well under the 500 NLOC limit:
- mod.rs (137L): struct, new(), dispatch_search, info
- search.rs (266L): dense/sparse/hybrid/batch/multi-query search
- query.rs (207L): VelesQL query/match/explain
- mutation.rs (202L): upsert/delete/flush/stream_insert + parse_payload DRY helper
- index.rs (81L): property/range index CRUD
